backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Pydantic
pydantic>=2.0.0
pydantic-settings>=2.0.0

# MikroTik API
librouteros>=3.2.0

# SSH
paramiko>=3.4.0

# WebSocket
websockets>=12.0

# HTTP client
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0

# =============================================================================
# New dependencies for extended features
# =============================================================================

# Authentication (JWT)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Scheduling
APScheduler>=3.10.0
croniter>=2.0.0
# cronstrue>=2.47.0  # Optional: Not available on all platforms

# Notifications
aiosmtplib>=3.0.0
# slack-sdk>=3.21.0  # Optional: Uncomment for Slack support
# python-telegram-bot>=21.0  # Optional: Uncomment for Telegram support

# Reports
reportlab>=4.0.0
openpyxl>=3.1.0
Jinja2>=3.1.0

# Monitoring
aioping>=0.4.0
